+++ wemake_python_styleguide/visitors/tokenize/wrong_comments.py
@@ -22,13 +22,17 @@ from wemake_python_styleguide.errors.tokens import WrongMagicCommentViolation
 from wemake_python_styleguide.visitors.base import BaseTokenVisitor
 
 NOQA_CHECK: Pattern = re.compile(r'^noqa:?($|[A-Z\d\,\s]+)')
+TYPE_CHECK: Pattern = re.compile(r'^type:\s?([\w\d\[\]\'\"\.]+)$')
 
 
 class WrongCommentVisitor(BaseTokenVisitor):
     """Checks comment tokens."""
 
+    def _get_comment_text(self, token: tokenize.TokenInfo) -> str:
+        return token.string[1:].strip()
+
     def _check_noqa(self, token: tokenize.TokenInfo) -> None:
-        comment_text = token.string[1:].strip()
+        comment_text = self._get_comment_text(token)
         match: Match = NOQA_CHECK.match(comment_text)
         if not match:
             return
@@ -36,9 +40,22 @@ class WrongCommentVisitor(BaseTokenVisitor):
         excludes = match.groups()[0].strip()
         if not excludes:
             # We can not pass the actual line here,
-            # since it will be ignored due to `noqa` comment:
+            # since it will be ignored due to `# noqa` comment:
             self.add_error(WrongMagicCommentViolation(text=comment_text))
+
+    def _check_typed_ast(self, token: tokenize.TokenInfo) -> None:
+        comment_text = self._get_comment_text(token)
+        match: Match = TYPE_CHECK.match(comment_text)
+        if not match:
+            return
+
+        declared_type = match.groups()[0].strip()
+        if declared_type != 'ignore':
+            self.add_error(
+                WrongMagicCommentViolation(token, text=comment_text),
+            )
 
     def visit_comment(self, token: tokenize.TokenInfo) -> None:
         """Performs comment checks."""
         self._check_noqa(token)
+        self._check_typed_ast(token)