+++ CHANGELOG.md
@@ -14,7 +14,7 @@ We used to have incremental versioning before `0.1.0`.
 - Adds `--i-control-code` option to ignore `InitModuleHasLogicViolation`
 - Adds check for underscored numbers
 - Forbids `u''` strings
-- Adds `flake8`, `noqa`, and `type` comments check
+- Adds `noqa` and `type` comments checks
 
 ### Misc